Generic classes DeleteDetailsContainerUseCase en SaveDetailsContainerUseCase Abstract generic class DetailsContainerViewModel - maar nog met beperkte functionaliteiten, eventueel later verder uitbreiden bij implementatie cinema
