+++ app/src/test/java/be/hogent/faith/faith/backpack/BackpackViewModelSaveAudioTest.kt
@@ -3,11 +3,12 @@ package be.hogent.faith.faith.backpack
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
 import be.hogent.faith.R
+import be.hogent.faith.domain.models.Backpack
 import be.hogent.faith.domain.models.User
 import be.hogent.faith.domain.models.detail.AudioDetail
 import be.hogent.faith.faith.backpackScreen.BackpackViewModel
 import be.hogent.faith.service.usecases.backpack.GetBackPackFilesDummyUseCase
-import be.hogent.faith.service.usecases.backpack.SaveBackpackDetailUseCase
+import be.hogent.faith.service.usecases.detailscontainer.SaveDetailsContainerDetailUseCase
 import io.mockk.Called
 import io.mockk.called
 import io.mockk.mockk
@@ -21,10 +22,10 @@ import org.junit.Assert.assertEquals
 
 class BackpackViewModelSaveAudioTest {
     private lateinit var viewModel: BackpackViewModel
-    private val saveAudioUseCase = mockk<SaveBackpackDetailUseCase>(relaxed = true)
+    private val saveAudioUseCase = mockk<SaveDetailsContainerDetailUseCase<Backpack>>(relaxed = true)
     private val getBackPackFilesDummyUseCase = mockk<GetBackPackFilesDummyUseCase>(relaxed = true)
     private val detail = mockk<AudioDetail>()
-    private val user: User = mockk()
+    private val user = mockk<User>(relaxed = true)
 
     @get:Rule
     val testRule = InstantTaskExecutorRule()
@@ -40,7 +41,7 @@ class BackpackViewModelSaveAudioTest {
 
     @Test
     fun backpackViewModel_saveAudio_callsUseCase() {
-        val params = slot<SaveBackpackDetailUseCase.Params>()
+        val params = slot<SaveDetailsContainerDetailUseCase.Params>()
 
         viewModel.saveAudioDetail(user, detail)
         verify { saveAudioUseCase.execute(capture(params), any()) }
@@ -55,7 +56,7 @@ class BackpackViewModelSaveAudioTest {
         val errorObserver = mockk<Observer<Int>>(relaxed = true)
         val successObserver = mockk<Observer<Int>>(relaxed = true)
         viewModel.errorMessage.observeForever(errorObserver)
-        viewModel.audioDetailSavedSuccessFully.observeForever(successObserver)
+        viewModel.infoMessage.observeForever(successObserver)
 
         // Act
         viewModel.saveAudioDetail(user, detail)
@@ -74,7 +75,7 @@ class BackpackViewModelSaveAudioTest {
         val errorObserver = mockk<Observer<Int>>(relaxed = true)
         val successObserver = mockk<Observer<Int>>(relaxed = true)
         viewModel.errorMessage.observeForever(errorObserver)
-        viewModel.audioDetailSavedSuccessFully.observeForever(successObserver)
+        viewModel.infoMessage.observeForever(successObserver)
 
         // Act
         viewModel.saveAudioDetail(user, detail)
