+++ app/src/test/java/be/hogent/faith/faith/backpack/BackpackViewModelSaveDrawingTest.kt
@@ -3,11 +3,12 @@ package be.hogent.faith.faith.backpack
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
 import be.hogent.faith.R
+import be.hogent.faith.domain.models.Backpack
 import be.hogent.faith.domain.models.User
 import be.hogent.faith.domain.models.detail.DrawingDetail
 import be.hogent.faith.faith.backpackScreen.BackpackViewModel
 import be.hogent.faith.service.usecases.backpack.GetBackPackFilesDummyUseCase
-import be.hogent.faith.service.usecases.backpack.SaveBackpackDetailUseCase
+import be.hogent.faith.service.usecases.detailscontainer.SaveDetailsContainerDetailUseCase
 import io.mockk.Called
 import io.mockk.called
 import io.mockk.mockk
@@ -22,10 +23,10 @@ import org.junit.Assert.assertEquals
 class BackpackViewModelSaveDrawingTest {
 
     private lateinit var viewModel: BackpackViewModel
-    private val saveDrawingUseCase = mockk<SaveBackpackDetailUseCase>(relaxed = true)
+    private val saveDrawingUseCase = mockk<SaveDetailsContainerDetailUseCase<Backpack>>(relaxed = true)
     private val getBackPackFilesDummyUseCase = mockk<GetBackPackFilesDummyUseCase>(relaxed = true)
     private val detail = mockk<DrawingDetail>()
-    private val user: User = mockk()
+    private val user = mockk<User>(relaxed = true)
 
     @get:Rule
     val testRule = InstantTaskExecutorRule()
@@ -41,7 +42,7 @@ class BackpackViewModelSaveDrawingTest {
 
     @Test
     fun backpackViewModel_saveDrawing_callsUseCase() {
-        val params = slot<SaveBackpackDetailUseCase.Params>()
+        val params = slot<SaveDetailsContainerDetailUseCase.Params>()
 
         viewModel.saveDrawingDetail(user, detail)
         verify { saveDrawingUseCase.execute(capture(params), any()) }
@@ -56,7 +57,7 @@ class BackpackViewModelSaveDrawingTest {
         val errorObserver = mockk<Observer<Int>>(relaxed = true)
         val successObserver = mockk<Observer<Int>>(relaxed = true)
         viewModel.errorMessage.observeForever(errorObserver)
-        viewModel.drawingDetailSavedSuccessFully.observeForever(successObserver)
+        viewModel.infoMessage.observeForever(successObserver)
 
         // Act
         viewModel.saveDrawingDetail(user, detail)
@@ -75,7 +76,7 @@ class BackpackViewModelSaveDrawingTest {
         val errorObserver = mockk<Observer<Int>>(relaxed = true)
         val successObserver = mockk<Observer<Int>>(relaxed = true)
         viewModel.errorMessage.observeForever(errorObserver)
-        viewModel.drawingDetailSavedSuccessFully.observeForever(successObserver)
+        viewModel.infoMessage.observeForever(successObserver)
 
         // Act
         viewModel.saveDrawingDetail(user, detail)
