+++ app/src/test/java/be/hogent/faith/faith/backpack/BackpackViewModelSaveExternalVideoTest.kt
@@ -14,17 +14,18 @@ import org.junit.Assert.assertEquals
 import androidx.lifecycle.Observer
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import be.hogent.faith.R
+import be.hogent.faith.domain.models.Backpack
 import be.hogent.faith.domain.models.User
 import be.hogent.faith.faith.backpackScreen.BackpackViewModel
 import be.hogent.faith.service.usecases.backpack.GetBackPackFilesDummyUseCase
-import be.hogent.faith.service.usecases.backpack.SaveBackpackDetailUseCase
+import be.hogent.faith.service.usecases.detailscontainer.SaveDetailsContainerDetailUseCase
 
 class BackpackViewModelSaveExternalVideoTest {
     private lateinit var viewModel: BackpackViewModel
-    private val saveExternalVideoUseCase = mockk<SaveBackpackDetailUseCase>(relaxed = true)
+    private val saveExternalVideoUseCase = mockk<SaveDetailsContainerDetailUseCase<Backpack>>(relaxed = true)
     private val getBackPackFilesDummyUseCase = mockk<GetBackPackFilesDummyUseCase>(relaxed = true)
     private val detail = mockk<ExternalVideoDetail>()
-    private val user = mockk<User>()
+    private val user = mockk<User>(relaxed = true)
 
     @get:Rule
     val testRule = InstantTaskExecutorRule()
@@ -40,7 +41,7 @@ class BackpackViewModelSaveExternalVideoTest {
 
     @Test
     fun backpackViewModel_saveExternalVideo_callsUseCase() {
-        val params = slot<SaveBackpackDetailUseCase.Params>()
+        val params = slot<SaveDetailsContainerDetailUseCase.Params>()
 
         viewModel.saveExternalVideoDetail(user, detail)
         verify { saveExternalVideoUseCase.execute(capture(params), any()) }
@@ -55,7 +56,7 @@ class BackpackViewModelSaveExternalVideoTest {
         val errorObserver = mockk<Observer<Int>>(relaxed = true)
         val successObserver = mockk<Observer<Int>>(relaxed = true)
         viewModel.errorMessage.observeForever(errorObserver)
-        viewModel.externalVideoSavedSuccessFully.observeForever(successObserver)
+        viewModel.infoMessage.observeForever(successObserver)
 
         // Act
         viewModel.saveExternalVideoDetail(user, detail)
@@ -74,7 +75,7 @@ class BackpackViewModelSaveExternalVideoTest {
         val errorObserver = mockk<Observer<Int>>(relaxed = true)
         val successObserver = mockk<Observer<Int>>(relaxed = true)
         viewModel.errorMessage.observeForever(errorObserver)
-        viewModel.photoDetailSavedSuccessFully.observeForever(successObserver)
+        viewModel.infoMessage.observeForever(successObserver)
 
         // Act
         viewModel.saveExternalVideoDetail(user, detail)
