+++ database/src/main/java/be/hogent/faith/database/firebase/FirebaseBackpackRepository.kt
@@ -5,8 +5,8 @@ import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 
 class FirebaseBackpackRepository(
-      fbAuth: FirebaseAuth,
-      firestore: FirebaseFirestore
+    fbAuth: FirebaseAuth,
+    firestore: FirebaseFirestore
 ) : FirebaseDetailContainerRepository<Backpack>(fbAuth, firestore) {
     override val containerName: String = "backpack"
 }
