+++ service/src/main/java/be/hogent/faith/service/usecases/di/ServiceModule.kt
@@ -1,5 +1,6 @@
 package be.hogent.faith.service.usecases.di
 
+import be.hogent.faith.domain.models.Backpack
 import be.hogent.faith.service.usecases.CreateUserUseCase
 import be.hogent.faith.service.usecases.GetUserUseCase
 import be.hogent.faith.service.usecases.IsUsernameUniqueUseCase
@@ -7,9 +8,8 @@ import be.hogent.faith.service.usecases.LoginUserUseCase
 import be.hogent.faith.service.usecases.LogoutUserUseCase
 import be.hogent.faith.service.usecases.RegisterUserUseCase
 import be.hogent.faith.service.usecases.detail.LoadDetailFileUseCase
-import be.hogent.faith.service.usecases.backpack.DeleteBackpackDetailUseCase
+import be.hogent.faith.service.usecases.detailscontainer.DeleteDetailsContainerDetailUseCase
 import be.hogent.faith.service.usecases.backpack.GetBackPackFilesDummyUseCase
-import be.hogent.faith.service.usecases.backpack.SaveBackpackDetailUseCase
 import be.hogent.faith.service.usecases.detail.audioDetail.CreateAudioDetailUseCase
 import be.hogent.faith.service.usecases.detail.drawingDetail.CreateDrawingDetailUseCase
 import be.hogent.faith.service.usecases.detail.drawingDetail.OverwriteDrawingDetailUseCase
@@ -18,6 +18,7 @@ import be.hogent.faith.service.usecases.detail.photoDetail.CreatePhotoDetailUseC
 import be.hogent.faith.service.usecases.detail.textDetail.CreateTextDetailUseCase
 import be.hogent.faith.service.usecases.detail.textDetail.LoadTextDetailUseCase
 import be.hogent.faith.service.usecases.detail.textDetail.OverwriteTextDetailUseCase
+import be.hogent.faith.service.usecases.detailscontainer.SaveDetailsContainerDetailUseCase
 import be.hogent.faith.service.usecases.event.SaveEventDetailUseCase
 import be.hogent.faith.service.usecases.event.GetEventFilesUseCase
 import be.hogent.faith.service.usecases.event.GetEventsUseCase
@@ -56,6 +57,11 @@ val serviceModule = module {
     factory { GetEventFilesUseCase(get(), get()) }
     factory { LoadDetailFileUseCase(get(), get()) }
     factory { GetBackPackFilesDummyUseCase(get(), get()) }
-    factory { SaveBackpackDetailUseCase(get(), get(), get()) }
-    factory { DeleteBackpackDetailUseCase(get(), get()) }
+    factory { SaveDetailsContainerDetailUseCase<Backpack>(get(), get(), get()) }
+    factory {
+        DeleteDetailsContainerDetailUseCase<Backpack>(
+            get(),
+            get()
+        )
+    }
 }
