+++ service/src/test/java/be/hogent/faith/service/usecases/backpack/SaveBackpackExternalVideoDetailUseCaseTest.kt
@@ -4,13 +4,14 @@ import be.hogent.faith.domain.models.Backpack
 import be.hogent.faith.domain.models.User
 import be.hogent.faith.domain.models.detail.ExternalVideoDetail
 import be.hogent.faith.domain.repository.DetailContainerRepository
+import be.hogent.faith.service.usecases.detailscontainer.SaveDetailsContainerDetailUseCase
 import be.hogent.faith.storage.IStorageRepository
 import io.mockk.mockk
 import io.reactivex.Scheduler
 import org.junit.Before
 
 class SaveBackpackExternalVideoDetailUseCaseTest {
-    private lateinit var saveBackpackExternalVideoDetailUseCase: SaveBackpackDetailUseCase
+    private lateinit var saveBackpackExternalVideoDetailUseCase: SaveDetailsContainerDetailUseCase<Backpack>
     private val scheduler: Scheduler = mockk()
     private val storageRepository: IStorageRepository = mockk(relaxed = true)
     private val repository: DetailContainerRepository<Backpack> = mockk(relaxed = true)
@@ -21,7 +22,7 @@ class SaveBackpackExternalVideoDetailUseCaseTest {
     @Before
     fun setUp() {
         saveBackpackExternalVideoDetailUseCase =
-                SaveBackpackDetailUseCase(
+            SaveDetailsContainerDetailUseCase(
                         repository,
                     storageRepository,
                         scheduler
