+++ service/src/test/java/be/hogent/faith/service/usecases/backpack/SaveBackpackTextDetailUseCaseTest.kt
@@ -4,6 +4,7 @@ import be.hogent.faith.domain.models.Backpack
 import be.hogent.faith.domain.models.User
 import be.hogent.faith.domain.models.detail.TextDetail
 import be.hogent.faith.domain.repository.DetailContainerRepository
+import be.hogent.faith.service.usecases.detailscontainer.SaveDetailsContainerDetailUseCase
 import be.hogent.faith.storage.IStorageRepository
 import io.mockk.every
 import io.mockk.mockk
@@ -16,7 +17,7 @@ import org.junit.Test
 
 class SaveBackpackTextDetailUseCaseTest {
 
-    private lateinit var saveBackpackTextDetailUseCase: SaveBackpackDetailUseCase
+    private lateinit var saveBackpackTextDetailUseCase: SaveDetailsContainerDetailUseCase<Backpack>
     private val scheduler: Scheduler = mockk()
     private val repository: DetailContainerRepository<Backpack> = mockk(relaxed = true)
     private val storageRepository: IStorageRepository = mockk(relaxed = true)
@@ -27,7 +28,7 @@ class SaveBackpackTextDetailUseCaseTest {
     @Before
     fun setUp() {
         saveBackpackTextDetailUseCase =
-            SaveBackpackDetailUseCase(
+            SaveDetailsContainerDetailUseCase(
                 repository,
                 storageRepository,
                 scheduler
@@ -39,7 +40,7 @@ class SaveBackpackTextDetailUseCaseTest {
         // Arrange
         every { repository.insertDetail(detail, user) } returns Maybe.fromSingle { detail }
         every { storageRepository.saveDetailFileForContainer(user.backpack, detail) } returns Single.just(detail)
-        val params = SaveBackpackDetailUseCase.Params(user, detail)
+        val params = SaveDetailsContainerDetailUseCase.Params(user, user.backpack, detail)
 
         // Act
         saveBackpackTextDetailUseCase.buildUseCaseObservable(params).test()
